requirements.txt
aiogram==3.4.1
fastapi==0.115.2
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.5.3
python-dotenv==1.0.1
numpy==1.26.4
vaderSentiment==3.3.2

